refactor: split duplication audit modules
