+++ src/core/code_audit/duplication.rs
@@ -23,7 +23,9 @@ use crate::core::component::DuplicationDetectorConfig;
 
 mod intra_method;
 
-use intra_method::is_low_information_literal_or_error_block;
+pub(crate) use intra_method::detect_intra_method_duplicates;
+#[cfg(test)]
+use intra_method::{has_logic_signal, is_scaffolding_line};
 
 /// Minimum number of locations for a function to count as duplicated.
 const MIN_DUPLICATE_LOCATIONS: usize = 2;
@@ -439,207 +441,6 @@ pub(crate) fn detect_near_duplicates(fingerprints: &[&FileFingerprint]) -> Vec<F
     findings
 }
 
-// ============================================================================
-// Intra-Method Duplication Detection
-// ============================================================================
-
-/// Minimum number of non-blank, non-comment lines for a block to be
-/// considered meaningful. Blocks shorter than this are too trivial to flag.
-const MIN_INTRA_BLOCK_LINES: usize = 5;
-
-/// Detect duplicated code blocks within the same method/function.
-///
-/// For each method in each file, extracts the method body from the file
-/// content and uses a sliding window of `MIN_INTRA_BLOCK_LINES` normalized
-/// lines. When the same window hash appears at two non-overlapping positions
-/// within one method, it means a block of code was copy-pasted (merge
-/// artifacts, copy-paste errors, etc.).
-pub(crate) fn detect_intra_method_duplicates(fingerprints: &[&FileFingerprint]) -> Vec<Finding> {
-    let mut findings = Vec::new();
-
-    for fp in fingerprints {
-        if fp.content.is_empty() {
-            continue;
-        }
-
-        let file_lines: Vec<&str> = fp.content.lines().collect();
-
-        for method_name in &fp.methods {
-            let Some((body_start, body_end)) = find_method_body(&file_lines, method_name) else {
-                continue;
-            };
-
-            // Extract body lines (excluding the opening/closing brace lines)
-            if body_start + 1 >= body_end {
-                continue;
-            }
-            let body_lines: Vec<&str> = file_lines[body_start + 1..body_end].to_vec();
-
-            if body_lines.len() < MIN_INTRA_BLOCK_LINES * 2 {
-                // Body too short to contain two meaningful duplicate blocks
-                continue;
-            }
-
-            // Build list of (original_body_index, normalized_text) for non-blank
-            // non-comment lines
-            let normalized: Vec<(usize, String)> = body_lines
-                .iter()
-                .enumerate()
-                .filter_map(|(i, line)| {
-                    let trimmed = line.trim();
-                    if trimmed.is_empty() || is_comment_only(trimmed) {
-                        None
-                    } else {
-                        Some((i, normalize_line(trimmed)))
-                    }
-                })
-                .collect();
-
-            if normalized.len() < MIN_INTRA_BLOCK_LINES * 2 {
-                continue;
-            }
-
-            // Hash each sliding window of MIN_INTRA_BLOCK_LINES consecutive
-            // normalized lines. Store (hash, start_body_idx, end_body_idx).
-            let mut window_hashes: Vec<(u64, usize, usize)> = Vec::new();
-
-            for win_start in 0..=normalized.len() - MIN_INTRA_BLOCK_LINES {
-                let win_end = win_start + MIN_INTRA_BLOCK_LINES;
-                let mut hasher = std::hash::DefaultHasher::new();
-                for (_, norm_line) in &normalized[win_start..win_end] {
-                    std::hash::Hash::hash(norm_line, &mut hasher);
-                }
-                let hash = std::hash::Hasher::finish(&hasher);
-
-                let orig_start = normalized[win_start].0;
-                let orig_end = normalized[win_end - 1].0;
-
-                window_hashes.push((hash, orig_start, orig_end));
-            }
-
-            // Group by hash, look for non-overlapping pairs
-            let mut hash_positions: HashMap<u64, Vec<(usize, usize)>> = HashMap::new();
-            for (hash, start, end) in &window_hashes {
-                hash_positions
-                    .entry(*hash)
-                    .or_default()
-                    .push((*start, *end));
-            }
-
-            let mut reported = false;
-            let mut suppressed_ranges: Vec<(usize, usize)> = Vec::new();
-
-            let mut duplicate_windows: Vec<&Vec<(usize, usize)>> =
-                hash_positions.values().collect();
-            duplicate_windows.sort_by_key(|positions| positions.first().copied());
-
-            for positions in duplicate_windows {
-                if reported || positions.len() < 2 {
-                    continue;
-                }
-
-                let first = positions[0];
-                for other in &positions[1..] {
-                    // Non-overlapping: second block starts after first block ends
-                    if other.0 <= first.1 {
-                        continue;
-                    }
-
-                    if is_inside_suppressed_range(first, &suppressed_ranges)
-                        || is_inside_suppressed_range(*other, &suppressed_ranges)
-                    {
-                        continue;
-                    }
-
-                    // Extend the match: keep sliding forward while lines match
-                    let first_norm_idx = normalized
-                        .iter()
-                        .position(|(i, _)| *i == first.0)
-                        .unwrap_or(0);
-                    let other_norm_idx = normalized
-                        .iter()
-                        .position(|(i, _)| *i == other.0)
-                        .unwrap_or(0);
-
-                    let mut match_len = MIN_INTRA_BLOCK_LINES;
-                    while first_norm_idx + match_len < normalized.len()
-                        && other_norm_idx + match_len < normalized.len()
-                        && first_norm_idx + match_len < other_norm_idx
-                    {
-                        if normalized[first_norm_idx + match_len].1
-                            == normalized[other_norm_idx + match_len].1
-                        {
-                            match_len += 1;
-                        } else {
-                            break;
-                        }
-                    }
-
-                    // Suppress structural-syntax-only windows. Match-arm tails
-                    // (`},`, `)?;`, `Ok((...))`, closing brace, bare-identifier
-                    // struct-literal fields) repeat naturally across sibling
-                    // dispatch branches in `run_*` functions — they're not
-                    // merge artifacts or copy-paste, they're Rust syntax.
-                    // A block is worth flagging only if it contains at least
-                    // one logic-bearing line.
-                    if is_structural_syntax_only(&normalized, first_norm_idx, match_len) {
-                        continue;
-                    }
-
-                    if is_branch_shape_repetition(&body_lines, first, *other, match_len)
-                        || is_low_information_literal_or_error_block(
-                            &normalized,
-                            first_norm_idx,
-                            match_len,
-                            MIN_INTRA_BLOCK_LINES,
-                        )
-                    {
-                        suppressed_ranges
-                            .push((first.0, normalized[first_norm_idx + match_len - 1].0));
-                        suppressed_ranges
-                            .push((other.0, normalized[other_norm_idx + match_len - 1].0));
-                        continue;
-                    }
-
-                    // Convert body-relative line numbers to 1-indexed file lines
-                    let first_file_line = body_start + 1 + first.0 + 1;
-                    let other_file_line = body_start + 1 + other.0 + 1;
-
-                    findings.push(Finding {
-                        convention: "intra-method-duplication".to_string(),
-                        severity: Severity::Warning,
-                        file: fp.relative_path.clone(),
-                        description: format!(
-                            "Duplicated block in `{}` — {} identical lines at line {} and line {}",
-                            method_name, match_len, first_file_line, other_file_line
-                        ),
-                        suggestion: format!(
-                            "Function `{}` contains a duplicated code block ({} lines). \
-                             This is often a merge artifact or copy-paste error. \
-                             Remove the duplicate or extract shared logic.",
-                            method_name, match_len
-                        ),
-                        kind: AuditFinding::IntraMethodDuplicate,
-                    });
-                    reported = true;
-                    break;
-                }
-
-                if reported {
-                    break;
-                }
-            }
-        }
-    }
-
-    findings.sort_by(|a, b| {
-        a.file
-            .cmp(&b.file)
-            .then_with(|| a.description.cmp(&b.description))
-    });
-    findings
-}
-
 /// Find the body of a method/function in the file lines.
 ///
 /// Returns `(open_brace_line, close_brace_line)` — the line indices of the
@@ -687,165 +488,6 @@ fn find_method_body(lines: &[&str], method_name: &str) -> Option<(usize, usize)>
     }
 
     None
-}
-
-/// Check if a line is comment-only (PHP, Rust, or shell style).
-fn is_comment_only(trimmed: &str) -> bool {
-    trimmed.starts_with("//")
-        || trimmed.starts_with("/*")
-        || trimmed.starts_with('*')
-        || trimmed.starts_with('#')
-}
-
-/// Normalize a line for hashing: collapse whitespace, lowercase.
-fn normalize_line(line: &str) -> String {
-    line.split_whitespace()
-        .collect::<Vec<_>>()
-        .join(" ")
-        .to_lowercase()
-}
-
-/// Return true when the window `normalized[start..start+len]` is pure
-/// syntactic scaffolding with no logic-bearing content.
-///
-/// A window is scaffolding when **every** line is one of:
-/// - pure punctuation closers (`}`, `},`, `)?;`, etc.)
-/// - a single identifier, optionally trailed by `,` (struct-literal or
-///   destructuring fields)
-/// - common match-arm glue (`=> {`, `} => {`)
-///
-/// **and** none of the lines contain logic signals (`=`, `let `, `if `,
-/// `for `, `while `, `match `, `return`, or a function-call shape
-/// `foo(` / `foo::bar(`). If a single line in the window carries any
-/// logic signal, the window is not scaffolding and gets flagged normally.
-///
-/// Match-arm tails (`)?;`, `Ok((x, 0))`, struct-literal closers) repeated
-/// across sibling arms of a dispatch `match` are structural noise, not
-/// duplication — this filter stops them from tripping the detector.
-fn is_structural_syntax_only(normalized: &[(usize, String)], start: usize, len: usize) -> bool {
-    let end = (start + len).min(normalized.len());
-    if start >= end {
-        return false;
-    }
-    let window = &normalized[start..end];
-
-    // If any line in the window looks logical, window is not scaffolding.
-    if window.iter().any(|(_, line)| has_logic_signal(line)) {
-        return false;
-    }
-
-    // Every line must match a known scaffolding shape.
-    window.iter().all(|(_, line)| is_scaffolding_line(line))
-}
-
-/// Lines that look like they do real work: assignment, control flow, or
-/// a user function call that isn't a dispatch-return wrapper.
-fn has_logic_signal(normalized: &str) -> bool {
-    let t = normalized.trim();
-
-    // Assignment or `let` binding.
-    if t.contains(" = ") || t.starts_with("let ") {
-        return true;
-    }
-
-    // Control flow keywords (normalized to lowercase by the caller).
-    for kw in ["if ", "for ", "while ", "match ", "return ", "loop ", "?;"] {
-        if t.contains(kw) && !matches!(t, ")?;" | "})?;") {
-            return true;
-        }
-    }
-
-    // Function / method calls that aren't bare dispatch-return wrappers.
-    // `ok(...)`, `err(...)`, `some(...)`, `none` by themselves are scaffolding
-    // (return-tail on a match arm); anything else with parens is real work.
-    if t.contains('(') {
-        let before_paren = t.split('(').next().unwrap_or("");
-        let head = before_paren.trim_end_matches(':').trim_end_matches(':');
-        let head = head.trim();
-        let is_return_wrapper = matches!(head, "ok" | "err" | "some")
-            || head.ends_with(" ok")
-            || head.ends_with(" err")
-            || head.ends_with(" some");
-        if !is_return_wrapper {
-            return true;
-        }
-    }
-
-    false
-}
-
-/// Does this normalized line match a known scaffolding shape?
-fn is_scaffolding_line(normalized: &str) -> bool {
-    let t = normalized.trim();
-    if t.is_empty() {
-        return true;
-    }
-
-    // Pure-punctuation closers: `}`, `},`, `)?;`, `))`, etc.
-    if t.chars()
-        .all(|c| matches!(c, '}' | ')' | '?' | ';' | ',' | '('))
-    {
-        return true;
-    }
-
-    // Bare identifier (optionally trailing comma) — struct-literal field or
-    // destructure.
-    let core = t.trim_end_matches(',');
-    if !core.is_empty() && core.chars().all(|c| c.is_ascii_alphanumeric() || c == '_') {
-        return true;
-    }
-
-    // Dispatch-return tails: `ok(...)`, `err(...)`, `some(...)`, `none`
-    // (optionally with trailing `?`, `;`, `,`).
-    let core = t.trim_end_matches([',', ';', '?']);
-    if core == "none"
-        || core.starts_with("ok(")
-        || core.starts_with("err(")
-        || core.starts_with("some(")
-    {
-        return true;
-    }
-
-    // Match-arm glue.
-    if t.ends_with("=> {") || t == "} => {" || t == "_ => {" {
-        return true;
-    }
-
-    false
-}
-
-/// Repeated blocks in sibling `if` / `else if` / `else` arms are usually local
-/// branch shape, not high-confidence copy/paste. Keep this deliberately narrow:
-/// long blocks can still indicate real duplication, and adjacent repeated logic
-/// outside branch arms is still reported.
-fn is_branch_shape_repetition(
-    body_lines: &[&str],
-    first: (usize, usize),
-    other: (usize, usize),
-    match_len: usize,
-) -> bool {
-    if match_len > 12 || first.1 >= other.0 || other.0 > body_lines.len() {
-        return false;
-    }
-
-    body_lines[first.1 + 1..other.0]
-        .iter()
-        .any(|line| is_branch_separator(line.trim()))
-}
-
-fn is_inside_suppressed_range(candidate: (usize, usize), ranges: &[(usize, usize)]) -> bool {
-    ranges
-        .iter()
-        .any(|(start, end)| candidate.0 >= *start && candidate.1 <= *end)
-}
-
-fn is_branch_separator(trimmed: &str) -> bool {
-    trimmed.starts_with("} else")
-        || trimmed.starts_with("else ")
-        || trimmed.starts_with("elseif ")
-        || trimmed.starts_with("} elseif")
-        || trimmed.ends_with("=> {")
-        || trimmed.starts_with("} => {")
 }
 
 // ============================================================================
@@ -1588,1798 +1230,4 @@ pub(crate) fn detect_parallel_implementations(
 // ============================================================================
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::code_audit::conventions::Language;
-
-    fn make_fingerprint(path: &str, methods: &[&str], hashes: &[(&str, &str)]) -> FileFingerprint {
-        make_fingerprint_with_structural(path, methods, hashes, &[])
-    }
-
-    fn make_fingerprint_with_structural(
-        path: &str,
-        methods: &[&str],
-        hashes: &[(&str, &str)],
-        structural: &[(&str, &str)],
-    ) -> FileFingerprint {
-        FileFingerprint {
-            relative_path: path.to_string(),
-            language: Language::Rust,
-            methods: methods.iter().map(|s| s.to_string()).collect(),
-            method_hashes: hashes
-                .iter()
-                .map(|(k, v)| (k.to_string(), v.to_string()))
-                .collect(),
-            structural_hashes: structural
-                .iter()
-                .map(|(k, v)| (k.to_string(), v.to_string()))
-                .collect(),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn detects_exact_duplicate() {
-        let fp1 = make_fingerprint("src/utils/io.rs", &["is_zero"], &[("is_zero", "abc123")]);
-        let fp2 = make_fingerprint(
-            "src/utils/validation.rs",
-            &["is_zero"],
-            &[("is_zero", "abc123")],
-        );
-
-        let findings = detect_duplicates(&[&fp1, &fp2], &std::collections::HashSet::new());
-
-        assert_eq!(findings.len(), 2, "Should emit one finding per location");
-        assert!(findings
-            .iter()
-            .all(|f| f.kind == AuditFinding::DuplicateFunction));
-        assert!(findings.iter().any(|f| f.file == "src/utils/io.rs"));
-        assert!(findings.iter().any(|f| f.file == "src/utils/validation.rs"));
-        assert!(findings[0].description.contains("is_zero"));
-    }
-
-    #[test]
-    fn duplicate_functions_under_tests_are_info_findings() {
-        let fp1 = make_fingerprint(
-            "tests/import/ability-smoke.php",
-            &["imp_assert"],
-            &[("imp_assert", "abc123")],
-        );
-        let fp2 = make_fingerprint(
-            "tests/import/adapter-smoke.php",
-            &["imp_assert"],
-            &[("imp_assert", "abc123")],
-        );
-
-        let findings = detect_duplicates(&[&fp1, &fp2], &std::collections::HashSet::new());
-
-        assert_eq!(findings.len(), 2);
-        assert!(findings
-            .iter()
-            .all(|finding| finding.severity == Severity::Info));
-        assert!(findings
-            .iter()
-            .all(|finding| finding.suggestion.contains("shared test helper")));
-    }
-
-    #[test]
-    fn no_duplicates_different_hashes() {
-        let fp1 = make_fingerprint("src/a.rs", &["process"], &[("process", "hash_a")]);
-        let fp2 = make_fingerprint("src/b.rs", &["process"], &[("process", "hash_b")]);
-
-        let findings = detect_duplicates(&[&fp1, &fp2], &std::collections::HashSet::new());
-        assert!(
-            findings.is_empty(),
-            "Different hashes should not flag duplicates"
-        );
-    }
-
-    #[test]
-    fn no_duplicates_single_location() {
-        let fp = make_fingerprint("src/only.rs", &["unique_fn"], &[("unique_fn", "abc123")]);
-
-        let findings = detect_duplicates(&[&fp], &std::collections::HashSet::new());
-        assert!(findings.is_empty(), "Single location is not a duplicate");
-    }
-
-    #[test]
-    fn three_way_duplicate() {
-        let fp1 = make_fingerprint("src/a.rs", &["helper"], &[("helper", "same_hash")]);
-        let fp2 = make_fingerprint("src/b.rs", &["helper"], &[("helper", "same_hash")]);
-        let fp3 = make_fingerprint("src/c.rs", &["helper"], &[("helper", "same_hash")]);
-
-        let findings = detect_duplicates(&[&fp1, &fp2, &fp3], &std::collections::HashSet::new());
-
-        assert_eq!(findings.len(), 3, "Should flag all 3 locations");
-        assert!(findings[0].suggestion.contains("3 files"));
-    }
-
-    #[test]
-    fn empty_method_hashes_no_findings() {
-        let fp1 = make_fingerprint("src/a.rs", &["foo", "bar"], &[]);
-        let fp2 = make_fingerprint("src/b.rs", &["foo", "bar"], &[]);
-
-        let findings = detect_duplicates(&[&fp1, &fp2], &std::collections::HashSet::new());
-        assert!(
-            findings.is_empty(),
-            "No hashes means no duplication findings"
-        );
-    }
-
-    #[test]
-    fn mixed_duplicates_and_unique() {
-        let fp1 = make_fingerprint(
-            "src/a.rs",
-            &["shared", "unique_a"],
-            &[("shared", "same"), ("unique_a", "hash_a")],
-        );
-        let fp2 = make_fingerprint(
-            "src/b.rs",
-            &["shared", "unique_b"],
-            &[("shared", "same"), ("unique_b", "hash_b")],
-        );
-
-        let findings = detect_duplicates(&[&fp1, &fp2], &std::collections::HashSet::new());
-
-        assert_eq!(findings.len(), 2, "Only 'shared' should be flagged");
-        assert!(findings.iter().all(|f| f.description.contains("shared")));
-    }
-
-    // ========================================================================
-    // DuplicateGroup / canonical selection tests
-    // ========================================================================
-
-    #[test]
-    fn group_picks_canonical_by_shortest_path() {
-        let fp1 = make_fingerprint("src/core/deep/nested/helper.rs", &["foo"], &[("foo", "h1")]);
-        let fp2 = make_fingerprint("src/utils.rs", &["foo"], &[("foo", "h1")]);
-
-        let groups = detect_duplicate_groups(&[&fp1, &fp2]);
-
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].canonical_file, "src/utils.rs");
-        assert_eq!(
-            groups[0].remove_from,
-            vec!["src/core/deep/nested/helper.rs"]
-        );
-    }
-
-    #[test]
-    fn group_prefers_utils_directory() {
-        let fp1 = make_fingerprint("src/core/a.rs", &["shared"], &[("shared", "h1")]);
-        let fp2 = make_fingerprint("src/utils/helpers.rs", &["shared"], &[("shared", "h1")]);
-
-        let groups = detect_duplicate_groups(&[&fp1, &fp2]);
-
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].canonical_file, "src/utils/helpers.rs");
-        assert_eq!(groups[0].remove_from, vec!["src/core/a.rs"]);
-    }
-
-    #[test]
-    fn group_alphabetical_tiebreaker() {
-        let fp1 = make_fingerprint("src/b.rs", &["dup"], &[("dup", "h1")]);
-        let fp2 = make_fingerprint("src/a.rs", &["dup"], &[("dup", "h1")]);
-
-        let groups = detect_duplicate_groups(&[&fp1, &fp2]);
-
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].canonical_file, "src/a.rs");
-    }
-
-    #[test]
-    fn group_three_way_has_two_removals() {
-        let fp1 = make_fingerprint("src/a.rs", &["f"], &[("f", "h")]);
-        let fp2 = make_fingerprint("src/b.rs", &["f"], &[("f", "h")]);
-        let fp3 = make_fingerprint("src/c.rs", &["f"], &[("f", "h")]);
-
-        let groups = detect_duplicate_groups(&[&fp1, &fp2, &fp3]);
-
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].remove_from.len(), 2);
-        assert!(!groups[0].remove_from.contains(&groups[0].canonical_file));
-    }
-
-    // ========================================================================
-    // Near-duplicate detection tests
-    // ========================================================================
-
-    /// Helper to build a fingerprint with content for body-line counting.
-    fn make_fp_with_content(
-        path: &str,
-        content: &str,
-        hashes: &[(&str, &str)],
-        structural: &[(&str, &str)],
-    ) -> FileFingerprint {
-        let mut fp = make_fingerprint_with_structural(path, &[], hashes, structural);
-        fp.content = content.to_string();
-        fp
-    }
-
-    #[test]
-    fn near_duplicate_detected_when_structural_match_but_exact_differs() {
-        // cache_path in two files: same structure, different constants.
-        // Use a 3-body-line shape so the function clears MIN_BODY_LINES
-        // (the trivial-body filter); the structural twins differ only by
-        // the constant referenced.
-        let content_a = "fn cache_path() -> Option<PathBuf> {\n    let base = paths::homeboy().ok()?;\n    let file = base.join(CACHE_A);\n    Some(file)\n}\n";
-        let content_b = "fn cache_path() -> Option<PathBuf> {\n    let base = paths::homeboy().ok()?;\n    let file = base.join(CACHE_B);\n    Some(file)\n}\n";
-
-        let fp1 = make_fp_with_content(
-            "src/core/update_check.rs",
-            content_a,
-            &[("cache_path", "hash_a")],
-            &[("cache_path", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/ext_update_check.rs",
-            content_b,
-            &[("cache_path", "hash_b")],
-            &[("cache_path", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-
-        assert_eq!(findings.len(), 2, "Should flag both locations");
-        assert!(findings
-            .iter()
-            .all(|f| f.kind == AuditFinding::NearDuplicate));
-        assert!(findings[0].description.contains("cache_path"));
-        assert_eq!(findings[0].severity, Severity::Info);
-    }
-
-    #[test]
-    fn near_duplicate_skips_exact_duplicates() {
-        // If exact hashes match, exact-duplicate detector already handles it
-        let fp1 = make_fingerprint_with_structural(
-            "src/a.rs",
-            &["helper"],
-            &[("helper", "SAME")],
-            &[("helper", "SAME_STRUCT")],
-        );
-        let fp2 = make_fingerprint_with_structural(
-            "src/b.rs",
-            &["helper"],
-            &[("helper", "SAME")],
-            &[("helper", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(findings.is_empty(), "Exact duplicates should be excluded");
-    }
-
-    #[test]
-    fn near_duplicate_skips_generic_names() {
-        let content = "fn run() {\n    do_something();\n    do_more();\n}\n";
-        let fp1 = make_fp_with_content(
-            "src/core/a.rs",
-            content,
-            &[("run", "hash_a")],
-            &[("run", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/b.rs",
-            content,
-            &[("run", "hash_b")],
-            &[("run", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(
-            findings.is_empty(),
-            "'run' is a generic name — should be skipped"
-        );
-    }
-
-    #[test]
-    fn near_duplicate_skips_command_core_pairs() {
-        let content = "fn deploy_site() {\n    connect();\n    upload();\n    verify();\n}\n";
-        let fp1 = make_fp_with_content(
-            "src/commands/deploy.rs",
-            content,
-            &[("deploy_site", "hash_a")],
-            &[("deploy_site", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/deploy.rs",
-            content,
-            &[("deploy_site", "hash_b")],
-            &[("deploy_site", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(findings.is_empty(), "Command↔core pair should be skipped");
-    }
-
-    #[test]
-    fn near_duplicate_skips_trivial_functions() {
-        // default_true is only 1 line — too trivial to refactor
-        let content = "fn default_true() -> bool { true }\n";
-        let fp1 = make_fp_with_content(
-            "src/core/defaults.rs",
-            content,
-            &[("default_true", "hash_a")],
-            &[("default_true", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/project.rs",
-            content,
-            &[("default_true", "hash_b")],
-            &[("default_true", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(findings.is_empty(), "Trivial functions should be skipped");
-    }
-
-    #[test]
-    fn near_duplicate_not_skipped_for_multi_line_core_functions() {
-        // Non-trivial functions in core/ (not commands/) SHOULD be flagged
-        let content = "fn cache_path() -> Option<PathBuf> {\n    let base = paths::homeboy()?;\n    let file = base.join(FILENAME);\n    Some(file)\n}\n";
-        let fp1 = make_fp_with_content(
-            "src/core/update.rs",
-            content,
-            &[("cache_path", "hash_a")],
-            &[("cache_path", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/ext_update.rs",
-            content,
-            &[("cache_path", "hash_b")],
-            &[("cache_path", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert_eq!(
-            findings.len(),
-            2,
-            "Non-trivial core↔core near-duplicates should be flagged"
-        );
-    }
-
-    #[test]
-    fn near_duplicate_skips_all_command_files() {
-        // Multiple command files with same structural hash — normal pattern
-        let content = "fn components() {\n    let list = config::list();\n    for item in list {\n        output::print(item);\n    }\n}\n";
-        let fp1 = make_fp_with_content(
-            "src/commands/fleet.rs",
-            content,
-            &[("components", "hash_a")],
-            &[("components", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/commands/project.rs",
-            content,
-            &[("components", "hash_b")],
-            &[("components", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(findings.is_empty(), "All-commands group should be skipped");
-    }
-
-    // ========================================================================
-    // count_body_lines — measures body lines strictly between braces (#1517)
-    // ========================================================================
-
-    #[test]
-    fn count_body_lines_zero_for_single_line_body() {
-        // `fn x() -> u32 { 0 }` — opening and closing brace on the same line.
-        // Zero lines strictly between them, so zero body lines.
-        let content = "fn x() -> u32 { 0 }\n";
-        let mut fp = make_fingerprint("src/x.rs", &["x"], &[]);
-        fp.content = content.to_string();
-
-        assert_eq!(
-            count_body_lines(&fp, "x"),
-            0,
-            "single-line body should report 0 body lines"
-        );
-    }
-
-    #[test]
-    fn count_body_lines_one_for_three_line_shape() {
-        // The standard formatter shape:
-        //   fn x() -> u32 {
-        //       0
-        //   }
-        // Exactly one line strictly between the braces.
-        let content = "fn x() -> u32 {\n    0\n}\n";
-        let mut fp = make_fingerprint("src/x.rs", &["x"], &[]);
-        fp.content = content.to_string();
-
-        assert_eq!(
-            count_body_lines(&fp, "x"),
-            1,
-            "three-line shape should report 1 body line"
-        );
-    }
-
-    #[test]
-    fn count_body_lines_counts_actual_body_statements() {
-        // Multi-line body with 4 statements between the braces.
-        let content = "fn process(items: &[Item]) -> Result {\n    let mut out = Vec::new();\n    for item in items {\n        out.push(item.clone());\n    }\n    Ok(out)\n}\n";
-        let mut fp = make_fingerprint("src/process.rs", &["process"], &[]);
-        fp.content = content.to_string();
-
-        // Lines strictly between `{` and `}`:
-        //   let mut out = Vec::new();
-        //   for item in items {
-        //       out.push(item.clone());
-        //   }
-        //   Ok(out)
-        // → 5 body lines.
-        assert_eq!(
-            count_body_lines(&fp, "process"),
-            5,
-            "should count actual body lines (5), not the wrapping span (7)"
-        );
-    }
-
-    #[test]
-    fn near_duplicate_skips_idiomatic_collection_methods() {
-        // The triggering case for #1517: every Vec/HashMap wrapper in the
-        // ecosystem defines `fn len(&self) -> usize { self.inner.len() }`,
-        // and Clippy's `len_without_is_empty` lint requires `is_empty`
-        // alongside it. Two structs each defining both methods must NOT
-        // produce near_duplicate findings.
-        let content_a = "struct A { inner: Vec<u8> }\nimpl A {\n    pub fn len(&self) -> usize {\n        self.inner.len()\n    }\n    pub fn is_empty(&self) -> bool {\n        self.inner.is_empty()\n    }\n}\n";
-        let content_b = "struct B { inner: HashMap<String, u32> }\nimpl B {\n    pub fn len(&self) -> usize {\n        self.inner.len()\n    }\n    pub fn is_empty(&self) -> bool {\n        self.inner.is_empty()\n    }\n}\n";
-
-        let fp1 = make_fp_with_content(
-            "src/core/a.rs",
-            content_a,
-            &[("len", "hash_a_len"), ("is_empty", "hash_a_emp")],
-            &[("len", "SAME_LEN"), ("is_empty", "SAME_EMP")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/b.rs",
-            content_b,
-            &[("len", "hash_b_len"), ("is_empty", "hash_b_emp")],
-            &[("len", "SAME_LEN"), ("is_empty", "SAME_EMP")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert!(
-            findings.is_empty(),
-            "idiomatic collection-wrapper methods (`len`, `is_empty`) must not be flagged as near-duplicates; got {} finding(s): {:?}",
-            findings.len(),
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>(),
-        );
-    }
-
-    #[test]
-    fn near_duplicate_still_flags_real_duplicates() {
-        // Regression guard against over-suppressing: a non-trivially-named
-        // method with identical structural hash but different body hashes
-        // across two files (and a 3+ body-line shape) MUST still be flagged.
-        let content_a = "fn compute_fixability(item: &Item) -> bool {\n    let score = item.score();\n    let threshold = THRESHOLD_A;\n    score > threshold\n}\n";
-        let content_b = "fn compute_fixability(item: &Item) -> bool {\n    let score = item.score();\n    let threshold = THRESHOLD_B;\n    score > threshold\n}\n";
-
-        let fp1 = make_fp_with_content(
-            "src/core/a.rs",
-            content_a,
-            &[("compute_fixability", "hash_a")],
-            &[("compute_fixability", "SAME_STRUCT")],
-        );
-        let fp2 = make_fp_with_content(
-            "src/core/b.rs",
-            content_b,
-            &[("compute_fixability", "hash_b")],
-            &[("compute_fixability", "SAME_STRUCT")],
-        );
-
-        let findings = detect_near_duplicates(&[&fp1, &fp2]);
-        assert_eq!(
-            findings.len(),
-            2,
-            "real near-duplicates (non-idiomatic name, multi-line body, distinct body hashes) must still be flagged",
-        );
-        assert!(findings
-            .iter()
-            .all(|f| f.description.contains("compute_fixability")));
-    }
-
-    // ========================================================================
-    // Intra-method duplication tests
-    // ========================================================================
-
-    #[test]
-    fn intra_method_detects_duplicated_block() {
-        // Simulate a merge artifact: same 5-line block appears twice
-        let content = "<?php\nclass PipelineSteps {\n    public function handle_update( $request ) {\n        $config = array();\n        $has_provider = $request->has_param( 'provider' );\n        $has_model = $request->has_param( 'model' );\n        $has_prompt = $request->has_param( 'system_prompt' );\n        $has_disabled = $request->has_param( 'disabled_tools' );\n        $has_key = $request->has_param( 'ai_api_key' );\n\n        if ( $has_provider ) {\n            $config['provider'] = sanitize_text_field( $request->get_param( 'provider' ) );\n        }\n\n        $has_provider = $request->has_param( 'provider' );\n        $has_model = $request->has_param( 'model' );\n        $has_prompt = $request->has_param( 'system_prompt' );\n        $has_disabled = $request->has_param( 'disabled_tools' );\n        $has_key = $request->has_param( 'ai_api_key' );\n\n        if ( $has_provider ) {\n            $config['provider'] = sanitize_text_field( $request->get_param( 'provider' ) );\n        }\n\n        return $config;\n    }\n}\n";
-
-        let mut fp = make_fingerprint(
-            "inc/Api/Pipelines/PipelineSteps.php",
-            &["handle_update"],
-            &[],
-        );
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-
-        assert!(
-            !findings.is_empty(),
-            "Should detect duplicated block within handle_update"
-        );
-        assert!(findings[0].kind == AuditFinding::IntraMethodDuplicate);
-        assert!(findings[0].description.contains("handle_update"));
-    }
-
-    #[test]
-    fn intra_method_no_false_positive_on_unique_code() {
-        let content = "<?php\nclass Handler {\n    public function process( $data ) {\n        $name = sanitize_text_field( $data['name'] );\n        $email = sanitize_email( $data['email'] );\n        $phone = sanitize_text_field( $data['phone'] );\n        $address = sanitize_text_field( $data['address'] );\n        $city = sanitize_text_field( $data['city'] );\n\n        $result = $this->save( $name, $email );\n        $this->notify( $result );\n        $this->log_action( $result );\n        $this->update_cache( $result );\n        $this->send_confirmation( $email );\n\n        return $result;\n    }\n}\n";
-
-        let mut fp = make_fingerprint("inc/Handler.php", &["process"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Unique code should not trigger intra-method duplication"
-        );
-    }
-
-    #[test]
-    fn intra_method_skips_short_methods() {
-        let content = "fn short() {\n    let a = 1;\n    let b = 2;\n    let c = a + b;\n    println!(\"{}\", c);\n}\n";
-
-        let mut fp = make_fingerprint("src/short.rs", &["short"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(findings.is_empty(), "Short methods should be skipped");
-    }
-
-    #[test]
-    fn intra_method_rust_function_duplicated_block() {
-        let content = "fn process_items(items: &[Item]) -> Vec<Result> {\n    let mut results = Vec::new();\n    let config = load_config();\n    let validator = Validator::new(&config);\n    let processor = Processor::new(&config);\n    let output = processor.run(&items[0]);\n\n    results.push(output);\n\n    let config = load_config();\n    let validator = Validator::new(&config);\n    let processor = Processor::new(&config);\n    let output = processor.run(&items[0]);\n\n    results.push(output);\n\n    results\n}\n";
-
-        let mut fp = make_fingerprint("src/core/pipeline.rs", &["process_items"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            !findings.is_empty(),
-            "Should detect duplicated block in Rust function"
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_match_arm_tail_scaffolding() {
-        // Sibling dispatch arms in a `run_*` match share a boilerplate tail:
-        //   )?;
-        //   Ok((Variant(output), 0))
-        //   }
-        //   OtherArm::Name { ... } => {
-        //
-        // After normalization these look like 5+ identical lines across arms,
-        // but they're Rust syntax, not duplicated logic. The scaffolding
-        // filter should suppress the finding.
-        //
-        // Each arm body here is intentionally one unique line plus the
-        // scaffolding tail — so the only thing that repeats is scaffolding.
-        let content = "\
-fn run_pr(args: PrArgs) -> Result {
-    match args.command {
-        PrCommand::Create {
-            comp_create,
-        } => {
-            do_create_thing(comp_create);
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-        PrCommand::Edit {
-            comp_edit,
-        } => {
-            do_edit_thing(comp_edit);
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-        PrCommand::Comment {
-            comp_comment,
-        } => {
-            do_comment_thing(comp_comment);
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-    }
-}
-";
-        let mut fp = make_fingerprint("src/commands/git.rs", &["run_pr"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Match-arm tail scaffolding should not be flagged as duplication; got {} finding(s): {:?}",
-            findings.len(),
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_still_flags_real_duplication_with_scaffolding_tails() {
-        // If the repeated block contains real logic (a `let` + a call that
-        // isn't an Ok/Err wrapper), we should still flag it even when it's
-        // surrounded by structural lines.
-        let content = "\
-fn process_twice() -> Result {
-    let items = load_items()?;
-    let validator = Validator::new();
-    let processor = Processor::new();
-    let output = processor.run(&items);
-    save_output(&output)?;
-
-    let items = load_items()?;
-    let validator = Validator::new();
-    let processor = Processor::new();
-    let output = processor.run(&items);
-    save_output(&output)?;
-
-    Ok(())
-}
-";
-        let mut fp = make_fingerprint("src/core/pipeline.rs", &["process_twice"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            !findings.is_empty(),
-            "Real duplication with logic lines should still be detected"
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_complementary_output_dto_tails() {
-        let content = r#"
-fn show(builtin: bool) -> CmdResult<ConfigOutput> {
-    if builtin {
-        Ok((
-            ConfigOutput {
-                command: "config.show".to_string(),
-                defaults: Some(defaults::builtin_defaults()),
-                config: None,
-                path: None,
-                exists: None,
-                pointer: None,
-                value: None,
-                deleted: None,
-            },
-            0,
-        ))
-    } else {
-        let config = defaults::load_config();
-        Ok((
-            ConfigOutput {
-                command: "config.show".to_string(),
-                config: Some(config),
-                defaults: None,
-                path: None,
-                exists: None,
-                pointer: None,
-                value: None,
-                deleted: None,
-            },
-            0,
-        ))
-    }
-}
-"#;
-        let mut fp = make_fingerprint("src/commands/config.rs", &["show"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Complementary DTO literal tails should not be flagged: {:?}",
-            findings
-                .iter()
-                .map(|f| f.description.as_str())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_repeated_error_envelopes() {
-        let content = r#"
-fn write_file_atomic(path: &Path, content: &str, operation: &str) -> Result<()> {
-    let parent = path.parent().ok_or_else(|| {
-        Error::internal_io(
-            format!("Invalid path: {}", path.display()),
-            Some(operation.to_string()),
-        )
-    })?;
-
-    let filename = path.file_name().ok_or_else(|| {
-        Error::internal_io(
-            format!("Invalid path: {}", path.display()),
-            Some(operation.to_string()),
-        )
-    })?;
-
-    let tmp_path = parent.join(format!("{}.tmp", filename.to_string_lossy()));
-    write_tmp(tmp_path, content)
-}
-"#;
-        let mut fp = make_fingerprint(
-            "src/core/engine/local_files.rs",
-            &["write_file_atomic"],
-            &[],
-        );
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Repeated error envelopes should not be flagged: {:?}",
-            findings
-                .iter()
-                .map(|f| f.description.as_str())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_short_sibling_branch_repetition() {
-        let content = r#"
-fn resolve_effective_glob(args: &Args, component: &Component) -> Result<Option<String>> {
-    if args.changed_only {
-        let changed_files = git::working_tree_changes(&component.local_path)?;
-        if changed_files.is_empty() {
-            println!("No files in working tree changes");
-            return Ok(Some(String::new()));
-        }
-
-        let abs_files: Vec<String> = changed_files
-            .iter()
-            .map(|f| format!("{}/{}", component.local_path, f))
-            .collect();
-
-        if abs_files.len() == 1 {
-            Ok(Some(abs_files[0].clone()))
-        } else {
-            Ok(Some(format!("{{{}}}", abs_files.join(","))))
-        }
-    } else if let Some(ref git_ref) = args.changed_since {
-        let changed_files = git::get_files_changed_since(&component.local_path, git_ref)?;
-        if changed_files.is_empty() {
-            println!("No files changed since {}", git_ref);
-            return Ok(Some(String::new()));
-        }
-
-        let abs_files: Vec<String> = changed_files
-            .iter()
-            .map(|f| format!("{}/{}", component.local_path, f))
-            .collect();
-
-        if abs_files.len() == 1 {
-            Ok(Some(abs_files[0].clone()))
-        } else {
-            Ok(Some(format!("{{{}}}", abs_files.join(","))))
-        }
-    } else {
-        Ok(args.glob.clone())
-    }
-}
-"#;
-        let mut fp = make_fingerprint(
-            "src/core/extension/lint/run.rs",
-            &["resolve_effective_glob"],
-            &[],
-        );
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Short sibling-branch repetition should not be flagged: {:?}",
-            findings
-                .iter()
-                .map(|f| f.description.as_str())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_repeated_multiline_call_argument_tails() {
-        let content = r#"
-fn env(extension: &Extension, local_path: &Path) -> Result<()> {
-    if let Some(detected) = run_component_env_detector(extension, local_path)? {
-        apply_component_env_detector_output(
-            detected,
-            &mut node_version,
-            &mut node_source,
-            &mut php_version,
-            &mut php_source,
-        );
-    }
-
-    if let Some(runtime) = extension.runtime.as_ref() {
-        apply_extension_runtime_requirements(
-            ext_id,
-            runtime,
-            &mut node_version,
-            &mut node_source,
-            &mut php_version,
-            &mut php_source,
-        );
-    }
-}
-"#;
-        let mut fp = make_fingerprint("src/commands/component.rs", &["env"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Repeated argument tails on different calls should not be flagged: {:?}",
-            findings
-                .iter()
-                .map(|f| f.description.as_str())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_ignores_repeated_match_arm_result_shapes() {
-        let content = r#"
-fn search(mode: SearchMode, line: &str, term: &str) {
-    match mode {
-        SearchMode::Boundary => {
-            for pos in find_boundary_matches(line, term) {
-                results.push(Match {
-                    file: relative.clone(),
-                    line: line_num + 1,
-                    column: pos + 1,
-                    matched: term.to_string(),
-                    context: line.to_string(),
-                });
-            }
-        }
-        SearchMode::Literal => {
-            for pos in find_literal_matches(line, term) {
-                results.push(Match {
-                    file: relative.clone(),
-                    line: line_num + 1,
-                    column: pos + 1,
-                    matched: term.to_string(),
-                    context: line.to_string(),
-                });
-            }
-        }
-    }
-}
-"#;
-        let mut fp = make_fingerprint("src/core/engine/codebase_scan.rs", &["search"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            findings.is_empty(),
-            "Repeated match-arm result shapes should not be flagged: {:?}",
-            findings
-                .iter()
-                .map(|f| f.description.as_str())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn intra_method_still_flags_adjacent_logic_copy_paste() {
-        let content = r#"
-fn rebuild_twice(items: &[Item]) -> Result<()> {
-    let config = load_config()?;
-    let validator = Validator::new(&config);
-    let processor = Processor::new(&config);
-    let output = processor.run(&items[0]);
-    save_output(&output)?;
-
-    let config = load_config()?;
-    let validator = Validator::new(&config);
-    let processor = Processor::new(&config);
-    let output = processor.run(&items[0]);
-    save_output(&output)?;
-
-    Ok(())
-}
-"#;
-        let mut fp = make_fingerprint("src/core/pipeline.rs", &["rebuild_twice"], &[]);
-        fp.content = content.to_string();
-
-        let findings = detect_intra_method_duplicates(&[&fp]);
-        assert!(
-            !findings.is_empty(),
-            "Adjacent repeated logic should still be reported"
-        );
-    }
-
-    #[test]
-    fn scaffolding_line_classifier() {
-        // Positive cases (structural).
-        for line in &[
-            "}",
-            "},",
-            ")",
-            ")?;",
-            "))",
-            "))?",
-            "path,",
-            "component_id,",
-            "path",
-            "ok((gitcommandoutput::pr(output), 0))",
-            "ok(output)",
-            "err(e)",
-            "none",
-            "} => {",
-            "_ => {",
-            "foo => {",
-        ] {
-            assert!(
-                is_scaffolding_line(line),
-                "Expected scaffolding: {:?}",
-                line
-            );
-        }
-
-        // Negative cases (real logic).
-        for line in &[
-            "let x = foo();",
-            "x = y + 1",
-            "if x.is_empty() {",
-            "for item in items {",
-            "compute(&items)?",
-            ".stdout(std::process::stdio::null())",
-        ] {
-            assert!(
-                !is_scaffolding_line(line) || has_logic_signal(line),
-                "Expected logic: {:?}",
-                line
-            );
-        }
-    }
-
-    #[test]
-    fn logic_signal_detector() {
-        assert!(has_logic_signal("let x = foo();"));
-        assert!(has_logic_signal("x = 1"));
-        assert!(has_logic_signal("if cond {"));
-        assert!(has_logic_signal("for x in y {"));
-        assert!(has_logic_signal("while true {"));
-        assert!(has_logic_signal("match thing {"));
-        assert!(has_logic_signal("return x"));
-        assert!(has_logic_signal(".stdout(something())"));
-        assert!(has_logic_signal("compute(&items)"));
-
-        // Return wrappers are NOT logic (they're structural tail expressions).
-        assert!(!has_logic_signal("ok(())"));
-        assert!(!has_logic_signal("ok((output, 0))"));
-        assert!(!has_logic_signal("err(e)"));
-        assert!(!has_logic_signal("some(x)"));
-        assert!(!has_logic_signal("none"));
-
-        // Pure punctuation is not logic.
-        assert!(!has_logic_signal("}"));
-        assert!(!has_logic_signal(")?;"));
-    }
-
-    #[test]
-    fn find_method_body_php() {
-        let content =
-            "<?php\nclass Foo {\n    public function bar() {\n        return 1;\n    }\n}\n";
-        let lines: Vec<&str> = content.lines().collect();
-        let result = find_method_body(&lines, "bar");
-        assert!(result.is_some());
-        let (open, close) = result.unwrap();
-        assert!(lines[open].contains('{'));
-        assert!(lines[close].contains('}'));
-    }
-
-    #[test]
-    fn find_method_body_rust() {
-        let content = "fn hello() {\n    println!(\"hi\");\n}\n";
-        let lines: Vec<&str> = content.lines().collect();
-        let result = find_method_body(&lines, "hello");
-        assert!(result.is_some());
-    }
-
-    #[test]
-    fn find_method_body_missing() {
-        let content = "fn other() {\n    println!(\"hi\");\n}\n";
-        let lines: Vec<&str> = content.lines().collect();
-        let result = find_method_body(&lines, "nonexistent");
-        assert!(result.is_none());
-    }
-
-    // ========================================================================
-    // Parallel Implementation Detection tests
-    // ========================================================================
-
-    fn make_fingerprint_with_content(
-        path: &str,
-        methods: &[&str],
-        content: &str,
-    ) -> FileFingerprint {
-        FileFingerprint {
-            relative_path: path.to_string(),
-            language: Language::Rust,
-            methods: methods.iter().map(|s| s.to_string()).collect(),
-            content: content.to_string(),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn detects_parallel_implementation() {
-        // Both bodies loop over a worklist — Looping ↔ Looping matches at the
-        // standard Jaccard floor. Mirrors the real `copy_dir_recursive` ↔
-        // `copy_directory` shape from issue #2334.
-        let fp1 = make_fingerprint_with_content(
-            "src/deploy.rs",
-            &["deploy_to_server"],
-            "fn deploy_to_server() {\n    for host in hosts {\n        validate_component();\n        build_artifact();\n        upload_to_host();\n        run_post_hooks();\n        notify_complete();\n    }\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/upgrade.rs",
-            &["upgrade_on_server"],
-            "fn upgrade_on_server() {\n    for host in hosts {\n        validate_component();\n        build_artifact();\n        upload_to_host();\n        run_post_hooks();\n        send_notification();\n    }\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(findings.len(), 2, "Should emit one finding per file");
-        assert!(findings
-            .iter()
-            .all(|f| f.kind == AuditFinding::ParallelImplementation));
-        assert!(findings.iter().any(|f| f.file == "src/deploy.rs"));
-        assert!(findings.iter().any(|f| f.file == "src/upgrade.rs"));
-    }
-
-    #[test]
-    fn no_parallel_for_unrelated_functions() {
-        let fp1 = make_fingerprint_with_content(
-            "src/deploy.rs",
-            &["deploy_to_server"],
-            "fn deploy_to_server() {\n    validate();\n    build();\n    upload();\n    notify();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/parser.rs",
-            &["parse_config"],
-            "fn parse_config() {\n    read_file();\n    tokenize();\n    parse_ast();\n    validate_schema();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        assert!(
-            findings.is_empty(),
-            "Completely different call sets should not flag"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_same_file() {
-        let fp = make_fingerprint_with_content(
-            "src/ops.rs",
-            &["deploy_op", "upgrade_op"],
-            "fn deploy_op() {\n    validate();\n    build();\n    upload();\n    notify();\n}\nfn upgrade_op() {\n    validate();\n    build();\n    upload();\n    notify();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&fp],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        assert!(
-            findings.is_empty(),
-            "Same-file methods should not be flagged as parallel"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_trivial_methods() {
-        let fp1 = make_fingerprint_with_content(
-            "src/a.rs",
-            &["small_a"],
-            "fn small_a() {\n    foo();\n    bar();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/b.rs",
-            &["small_b"],
-            "fn small_b() {\n    foo();\n    bar();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        assert!(
-            findings.is_empty(),
-            "Methods with < MIN_CALL_COUNT calls should be skipped"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_plumbing_only_call_patterns() {
-        let fs_helper = make_fingerprint_with_content(
-            "src/files.rs",
-            &["plugin_header_version"],
-            "fn plugin_header_version() {\n    path();\n    read_dir();\n    to_str();\n    success();\n}",
-        );
-        let extension_scan = make_fingerprint_with_content(
-            "src/extensions.rs",
-            &["scan_available_extensions"],
-            "fn scan_available_extensions() {\n    path();\n    read_dir();\n    to_str();\n    is_dir();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&fs_helper, &extension_scan],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Plumbing-only filesystem call overlap should not flag"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_command_wrapper_plumbing() {
-        let command_runner = make_fingerprint_with_content(
-            "src/command.rs",
-            &["succeeded_in"],
-            "fn succeeded_in() {\n    args();\n    current_dir();\n    output();\n    success();\n}",
-        );
-        let branch_reader = make_fingerprint_with_content(
-            "src/stack.rs",
-            &["current_branch"],
-            "fn current_branch() {\n    args();\n    current_dir();\n    output();\n    success();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&command_runner, &branch_reader],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Shared Command setup/result checks are plumbing, not a workflow"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_text_parsing_plumbing() {
-        let http_handler = make_fingerprint_with_content(
-            "src/core/daemon.rs",
-            &["handle_connection"],
-            "fn handle_connection() {\n    request.lines().next().split_whitespace();\n    route();\n    write_response();\n}",
-        );
-        let process_probe = make_fingerprint_with_content(
-            "src/core/server/client.rs",
-            &["probe_child_resources"],
-            "fn probe_child_resources() {\n    stdout.lines().next().split_whitespace();\n    parse_rss();\n    parse_cpu();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&http_handler, &process_probe],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Shared line tokenization is parsing plumbing, not a reusable workflow"
-        );
-    }
-
-    #[test]
-    fn no_parallel_for_deploy_plumbing_only_patterns() {
-        let artifact_deploy = make_fingerprint_with_content(
-            "src/core/deploy/safety_and_artifact.rs",
-            &["deploy_artifact"],
-            "fn deploy_artifact() {\n    quote_path();\n    execute();\n    failure();\n    render_map();\n    fix_deployed_permissions();\n}",
-        );
-        let override_deploy = make_fingerprint_with_content(
-            "src/core/deploy/version_overrides.rs",
-            &["deploy_with_override"],
-            "fn deploy_with_override() {\n    quote_path();\n    execute();\n    failure();\n    render_map();\n    fix_deployed_permissions();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&artifact_deploy, &override_deploy],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Shared SSH/deploy epilogue calls should not imply an extractable workflow"
-        );
-    }
-
-    #[test]
-    fn detects_parallel_implementation_after_plumbing_filter() {
-        // Both bodies loop over their PR list — Looping ↔ Looping clears the
-        // body-shape gate at the standard Jaccard floor.
-        let apply = make_fingerprint_with_content(
-            "src/core/stack/apply.rs",
-            &["apply_stack"],
-            "fn apply_stack() {\n    for pr in prs {\n        ensure_head_remote();\n        checkout_force();\n        fetch_pr_meta();\n        cherry_pick();\n        record_applied_pr();\n        run_git();\n        success();\n    }\n}",
-        );
-        let sync = make_fingerprint_with_content(
-            "src/core/stack/sync.rs",
-            &["sync_stack"],
-            "fn sync_stack() {\n    for pr in prs {\n        ensure_head_remote();\n        checkout_force();\n        fetch_pr_meta();\n        cherry_pick();\n        record_synced_pr();\n        run_git();\n        success();\n    }\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&apply, &sync],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(
-            findings.len(),
-            2,
-            "Domain-heavy stack pairs should still flag"
-        );
-        assert!(findings
-            .iter()
-            .any(|finding| finding.description.contains("`ensure_head_remote`")));
-        assert!(findings
-            .iter()
-            .all(|finding| !finding.description.contains("`run_git`")));
-    }
-
-    #[test]
-    fn no_parallel_for_generic_names() {
-        // "run" is in GENERIC_NAMES
-        let fp1 = make_fingerprint_with_content(
-            "src/a.rs",
-            &["run"],
-            "fn run() {\n    validate();\n    build();\n    upload();\n    notify();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/b.rs",
-            &["execute"],
-            "fn execute() {\n    validate();\n    build();\n    upload();\n    notify();\n}",
-        );
-
-        // "run" is skipped, so only one method in the pool — no pair to compare
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        // Only fp2's "execute" has a valid call sequence; fp1's "run" is filtered
-        // So there's only 1 candidate, no pair → no findings
-        assert!(findings.is_empty(), "Generic names should be filtered out");
-    }
-
-    #[test]
-    fn extract_calls_skips_keywords() {
-        let body = "if something() {\n    let x = process();\n    for item in list() {\n        handle(item);\n    }\n}";
-        let calls = extract_calls_from_body(body, &std::collections::HashSet::new());
-        assert!(calls.contains(&"something".to_string()));
-        assert!(calls.contains(&"process".to_string()));
-        assert!(calls.contains(&"list".to_string()));
-        assert!(calls.contains(&"handle".to_string()));
-        assert!(!calls.contains(&"if".to_string()));
-        assert!(!calls.contains(&"for".to_string()));
-        assert!(!calls.contains(&"let".to_string()));
-    }
-
-    #[test]
-    fn jaccard_identical_sets() {
-        let a = vec!["foo".to_string(), "bar".to_string()];
-        assert!((jaccard_similarity(&a, &a) - 1.0).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn jaccard_disjoint_sets() {
-        let a = vec!["foo".to_string()];
-        let b = vec!["bar".to_string()];
-        assert!((jaccard_similarity(&a, &b)).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn lcs_identical_sequences() {
-        let a = vec!["a".to_string(), "b".to_string(), "c".to_string()];
-        assert_eq!(lcs_length(&a, &a), 3);
-        assert!((lcs_ratio(&a, &a) - 1.0).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn lcs_partial_overlap() {
-        let a = vec!["a".to_string(), "b".to_string(), "c".to_string()];
-        let b = vec!["a".to_string(), "x".to_string(), "c".to_string()];
-        assert_eq!(lcs_length(&a, &b), 2); // a, c
-    }
-
-    #[test]
-    fn convention_methods_skip_parallel_detection() {
-        // Two methods with identical call patterns — would normally flag.
-        // Wrapped in a loop so they clear the body-shape gate.
-        let fp1 = make_fingerprint_with_content(
-            "src/deploy.rs",
-            &["registerAbilities"],
-            "fn registerAbilities() {\n    for ability in abilities {\n        validate_component();\n        build_artifact();\n        upload_to_host();\n        run_post_hooks();\n        notify_complete();\n    }\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/upgrade.rs",
-            &["registerAbility"],
-            "fn registerAbility() {\n    for ability in abilities {\n        validate_component();\n        build_artifact();\n        upload_to_host();\n        run_post_hooks();\n        send_notification();\n    }\n}",
-        );
-
-        // Without convention methods: flagged
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        assert_eq!(findings.len(), 2, "Should flag without convention context");
-
-        // With EITHER method as convention-expected: NOT flagged
-        let conv_methods: std::collections::HashSet<String> = ["registerAbilities"] // only one of the two
-            .iter()
-            .map(|s| s.to_string())
-            .collect();
-        let findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &conv_methods,
-            &DuplicationDetectorConfig::default(),
-        );
-        assert!(
-            findings.is_empty(),
-            "Pairs involving convention methods should not be flagged, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    // ========================================================================
-    // Body-shape gate tests (issue #2334)
-    // ========================================================================
-
-    #[test]
-    fn body_shape_gate_two_loops_with_shared_calls_flag() {
-        // Mirrors the real `copy_dir_recursive` ↔ `copy_directory` shape that
-        // we MUST keep flagging after the body-shape gate ships.
-        let copy_dir_recursive = make_fingerprint_with_content(
-            "src/core/extension/lifecycle.rs",
-            &["copy_dir_recursive"],
-            "fn copy_dir_recursive() {\n    create_dir_all(dst);\n    for entry in read_dir(src) {\n        copy_file_entry();\n        record_copied();\n        verify_target();\n    }\n}",
-        );
-        let copy_directory = make_fingerprint_with_content(
-            "src/core/engine/invocation.rs",
-            &["copy_directory"],
-            "fn copy_directory() {\n    create_dir_all(dst);\n    for entry in read_dir(src) {\n        copy_file_entry();\n        record_copied();\n        preserve_artifact();\n    }\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&copy_dir_recursive, &copy_directory],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(
-            findings.len(),
-            2,
-            "Two looping copy helpers with shared calls must still flag — that is the real finding from #2334"
-        );
-    }
-
-    #[test]
-    fn body_shape_gate_kills_single_file_vs_recursive_walk_fp() {
-        // The canonical FP from issue #2334:
-        // `copy_artifact_file` is StraightLine (single `fs::copy` after a
-        // `create_dir_all` of the parent), `copy_dir_recursive` is
-        // Looping+Recursive (recursive walk over `read_dir`). They share
-        // `create_dir_all` and `copy` but the workflows are not the same.
-        let copy_artifact_file = make_fingerprint_with_content(
-            "src/core/observation/store.rs",
-            &["copy_artifact_file"],
-            "fn copy_artifact_file() {\n    let parent = target_parent();\n    create_dir_all(parent);\n    copy(source, target);\n    verify_size();\n    record_copy();\n}",
-        );
-        let copy_dir_recursive = make_fingerprint_with_content(
-            "src/core/extension/lifecycle.rs",
-            &["copy_dir_recursive"],
-            "fn copy_dir_recursive() {\n    create_dir_all(dst);\n    for entry in read_dir(src) {\n        copy(entry, target);\n        verify_size();\n        record_copy();\n        copy_dir_recursive(entry, dst);\n    }\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&copy_artifact_file, &copy_dir_recursive],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Single-file copy (StraightLine) vs recursive walk (Looping) must not flag — got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn body_shape_gate_two_straight_line_below_raised_jaccard_floor() {
-        // Two StraightLine bodies: 4 shared calls, 6 union → Jaccard 0.667.
-        // Below the raised StraightLine floor of 0.7 → must NOT flag.
-        let helper_a = make_fingerprint_with_content(
-            "src/core/a.rs",
-            &["build_thing_a"],
-            "fn build_thing_a() {\n    let x = open_resource();\n    register_handler();\n    configure_options();\n    finalize_build();\n    emit_metric_a();\n}",
-        );
-        let helper_b = make_fingerprint_with_content(
-            "src/core/b.rs",
-            &["build_thing_b"],
-            "fn build_thing_b() {\n    let x = open_resource();\n    register_handler();\n    configure_options();\n    finalize_build();\n    emit_metric_b();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&helper_a, &helper_b],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Two StraightLine bodies at Jaccard 0.667 must not flag under the raised floor — got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn body_shape_gate_two_straight_line_above_raised_jaccard_floor() {
-        // Same StraightLine pair but with identical signal calls (Jaccard 1.0)
-        // — clears the raised floor and MUST flag. This proves the gate is a
-        // shape filter, not a blanket ban on StraightLine pairs.
-        let helper_a = make_fingerprint_with_content(
-            "src/core/a.rs",
-            &["build_thing_a"],
-            "fn build_thing_a() {\n    let x = open_resource();\n    register_handler();\n    configure_options();\n    finalize_build();\n    emit_metric();\n}",
-        );
-        let helper_b = make_fingerprint_with_content(
-            "src/core/b.rs",
-            &["build_thing_b"],
-            "fn build_thing_b() {\n    let x = open_resource();\n    register_handler();\n    configure_options();\n    finalize_build();\n    emit_metric();\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&helper_a, &helper_b],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(
-            findings.len(),
-            2,
-            "Two StraightLine bodies above the raised Jaccard floor must still flag"
-        );
-    }
-
-    #[test]
-    fn body_shape_gate_recursive_to_recursive_flags() {
-        // Two recursive helpers (no loop, but each calls itself) share the
-        // same workflow — Recursive ↔ Recursive is compatible and uses the
-        // standard Jaccard floor.
-        let walk_a = make_fingerprint_with_content(
-            "src/core/a.rs",
-            &["walk_tree_a"],
-            "fn walk_tree_a(node) {\n    visit_node();\n    record_step();\n    sanitize_value();\n    log_progress();\n    walk_tree_a(child);\n}",
-        );
-        let walk_b = make_fingerprint_with_content(
-            "src/core/b.rs",
-            &["walk_tree_b"],
-            "fn walk_tree_b(node) {\n    visit_node();\n    record_step();\n    sanitize_value();\n    log_progress();\n    walk_tree_b(child);\n}",
-        );
-
-        let findings = detect_parallel_implementations(
-            &[&walk_a, &walk_b],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(
-            findings.len(),
-            2,
-            "Recursive ↔ Recursive helpers with shared call set must flag"
-        );
-    }
-
-    // ========================================================================
-    // Extension-supplied trivial/plumbing call list tests (#2333)
-    // ========================================================================
-
-    #[test]
-    fn extension_trivial_calls_filter_out_of_signal() {
-        // Two parallel deploy/upgrade workflows that share several
-        // domain-meaningful calls — flagged by default. With `custom_helper`
-        // declared trivial via extension, that call is filtered out of the
-        // recorded sequence; the remaining shared signal is unchanged so
-        // this test specifically demonstrates the trivial path is wired
-        // (compare against the default-config sanity below).
-        let fp1 = make_fingerprint_with_content(
-            "src/deploy.rs",
-            &["deploy_to_server"],
-            "fn deploy_to_server() {\n    custom_helper();\n    validate_component();\n    build_artifact();\n    upload_to_host();\n    run_post_hooks();\n    notify_complete();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/upgrade.rs",
-            &["upgrade_on_server"],
-            "fn upgrade_on_server() {\n    custom_helper();\n    validate_component();\n    build_artifact();\n    upload_to_host();\n    run_post_hooks();\n    send_notification();\n}",
-        );
-
-        // Default config: flagged (has the shared workflow signal).
-        let default_findings = detect_parallel_implementations(
-            &[&fp1, &fp2],
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-        assert!(
-            !default_findings.is_empty(),
-            "Sanity: workflow pair must flag without extension config"
-        );
-        assert!(
-            default_findings
-                .iter()
-                .any(|f| f.description.contains("`custom_helper`")),
-            "Sanity: without trivial-list filtering, `custom_helper` should appear in the shared-call summary"
-        );
-
-        // Extension-supplied trivial removes `custom_helper` from sequences.
-        let cfg = DuplicationDetectorConfig {
-            trivial_calls: vec!["custom_helper".to_string()],
-            plumbing_calls: vec![],
-        };
-        let findings =
-            detect_parallel_implementations(&[&fp1, &fp2], &std::collections::HashSet::new(), &cfg);
-        // The pair still flags (other workflow signal remains), but the
-        // extension-trivial name is gone from the description.
-        assert!(
-            findings
-                .iter()
-                .all(|f| !f.description.contains("`custom_helper`")),
-            "Extension-trivial `custom_helper` must be filtered out of the call sequence and shared-call summary, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn body_shape_detection_smoke() {
-        assert_eq!(
-            detect_body_shape("    foo();\n    bar();\n", "thing"),
-            BodyShape::StraightLine
-        );
-        assert_eq!(
-            detect_body_shape("    for entry in items {\n        foo();\n    }\n", "thing"),
-            BodyShape::Looping
-        );
-        assert_eq!(
-            detect_body_shape("    items.iter().map(|x| x).collect();\n", "thing"),
-            BodyShape::Looping
-        );
-        assert_eq!(
-            detect_body_shape("    foo();\n    walk(child);\n", "walk"),
-            BodyShape::Recursive
-        );
-        // Identifier guard — `redo_walk` must not register as a self-call to `walk`.
-        assert_eq!(
-            detect_body_shape("    redo_walk(x);\n", "walk"),
-            BodyShape::StraightLine
-        );
-        // `format!` contains the substring `for` but must not register as Looping.
-        assert_eq!(
-            detect_body_shape("    let s = format!(\"x\");\n    foo();\n", "thing"),
-            BodyShape::StraightLine
-        );
-    }
-
-    #[test]
-    fn extension_plumbing_calls_filter_out_of_signal() {
-        // Two methods share only `log_event` as workflow overlap — everything
-        // else is unique. With default config, `log_event` is workflow signal
-        // and the pair flags. With `log_event` declared plumbing via extension,
-        // the shared signal collapses below MIN_CALL_COUNT and the pair is
-        // dropped.
-        let fp1 = make_fingerprint_with_content(
-            "src/a.rs",
-            &["worker_a"],
-            "fn worker_a() {\n    log_event();\n    log_event();\n    log_event();\n    log_event();\n    step_a1();\n    step_a2();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/b.rs",
-            &["worker_b"],
-            "fn worker_b() {\n    log_event();\n    log_event();\n    log_event();\n    log_event();\n    step_b1();\n    step_b2();\n}",
-        );
-
-        // Without extension config, log_event is recorded as signal — but with
-        // the existing built-in idiomatic floors, results vary. We only assert
-        // the extension hook genuinely silences any pairing.
-        let cfg = DuplicationDetectorConfig {
-            trivial_calls: vec![],
-            plumbing_calls: vec!["log_event".to_string()],
-        };
-        let findings =
-            detect_parallel_implementations(&[&fp1, &fp2], &std::collections::HashSet::new(), &cfg);
-        assert!(
-            findings
-                .iter()
-                .all(|f| !f.description.contains("`log_event`")),
-            "Extension-plumbing `log_event` must be removed from signal calls / shared-call summary, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn extension_call_lists_fix_env_path_helper_fp_2333() {
-        // Direct regression for issue #2333: `cache_fallback_root` ↔ `homeboy_data`
-        // both call `var`, `cfg`, `not`, `internal_unexpected` (Rust env-derived
-        // path plumbing). Without extension hints the detector flags them; with
-        // the rust manifest declaring those calls as trivial/plumbing, no FP.
-        let cache = make_fingerprint_with_content(
-            "src/core/cache.rs",
-            &["cache_fallback_root"],
-            "fn cache_fallback_root() {\n    var();\n    cfg();\n    not();\n    internal_unexpected();\n    join_cache_path();\n}",
-        );
-        let data = make_fingerprint_with_content(
-            "src/core/data.rs",
-            &["homeboy_data"],
-            "fn homeboy_data() {\n    var();\n    cfg();\n    not();\n    internal_unexpected();\n    join_data_path();\n}",
-        );
-
-        // Default config: detector may still flag (this is the FP we are fixing).
-        // We do NOT assert a specific shape here — issue #2334 covers a body-shape
-        // gate that may also suppress this. We only require that the extension
-        // hook genuinely silences it.
-
-        // With extension config matching the rust manifest:
-        let cfg = DuplicationDetectorConfig {
-            trivial_calls: vec!["var".to_string(), "cfg".to_string(), "not".to_string()],
-            plumbing_calls: vec!["internal_unexpected".to_string()],
-        };
-        let findings = detect_parallel_implementations(
-            &[&cache, &data],
-            &std::collections::HashSet::new(),
-            &cfg,
-        );
-        assert!(
-            findings.is_empty(),
-            "Issue #2333: env-derived path helpers should NOT flag once rust extension supplies trivial/plumbing lists, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn extension_lists_augment_built_in_floor_not_replace() {
-        // Built-in floors must remain active even when an extension supplies
-        // its own lists. Two methods sharing only built-in trivial calls
-        // (`to_string`, `clone`, `unwrap`) should not flag, regardless of
-        // extension config contents.
-        let fp1 = make_fingerprint_with_content(
-            "src/a.rs",
-            &["render_a"],
-            "fn render_a() {\n    to_string();\n    clone();\n    unwrap();\n    len();\n    iter();\n}",
-        );
-        let fp2 = make_fingerprint_with_content(
-            "src/b.rs",
-            &["render_b"],
-            "fn render_b() {\n    to_string();\n    clone();\n    unwrap();\n    len();\n    iter();\n}",
-        );
-
-        // Extension config that does NOT mention to_string/clone/etc.
-        let cfg = DuplicationDetectorConfig {
-            trivial_calls: vec!["something_unrelated".to_string()],
-            plumbing_calls: vec!["another_unrelated".to_string()],
-        };
-        let findings =
-            detect_parallel_implementations(&[&fp1, &fp2], &std::collections::HashSet::new(), &cfg);
-        assert!(
-            findings.is_empty(),
-            "Built-in trivial floor must remain active even with extension config, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn corpus_common_calls_filter_broad_boilerplate_signal() {
-        // Generic regression for issue #2398: if a call tuple appears across a
-        // broad slice of the scanned component, it is scaffolding for this
-        // detector. The names are intentionally framework-neutral fixtures;
-        // extensions can still provide explicit trivial/plumbing lists.
-        let mut fingerprints = Vec::new();
-
-        for idx in 0..8 {
-            let method = format!("boilerplate_holder_{idx}");
-            let content = format!(
-                "fn {method}() {{\n    scaffold_response();\n    read_request();\n    default_payload();\n    validate_presence();\n    filler_{idx}();\n}}"
-            );
-            fingerprints.push(make_fingerprint_with_content(
-                &format!("src/common_{idx}.rs"),
-                &[method.as_str()],
-                &content,
-            ));
-        }
-
-        let first = make_fingerprint_with_content(
-            "src/a.rs",
-            &["create_item"],
-            "fn create_item() {\n    scaffold_response();\n    read_request();\n    default_payload();\n    validate_presence();\n    create_specific_step();\n}",
-        );
-        let second = make_fingerprint_with_content(
-            "src/b.rs",
-            &["delete_item"],
-            "fn delete_item() {\n    scaffold_response();\n    read_request();\n    default_payload();\n    validate_presence();\n    delete_specific_step();\n}",
-        );
-        fingerprints.push(first);
-        fingerprints.push(second);
-
-        let refs = fingerprints.iter().collect::<Vec<_>>();
-        let findings = detect_parallel_implementations(
-            &refs,
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert!(
-            findings.is_empty(),
-            "Corpus-common scaffolding calls should not produce a parallel implementation finding, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn corpus_common_calls_preserve_domain_specific_signal() {
-        let mut fingerprints = Vec::new();
-
-        for idx in 0..8 {
-            let method = format!("boilerplate_holder_{idx}");
-            let content = format!(
-                "fn {method}() {{\n    scaffold_response();\n    read_request();\n    default_payload();\n    validate_presence();\n    filler_{idx}();\n}}"
-            );
-            fingerprints.push(make_fingerprint_with_content(
-                &format!("src/common_{idx}.rs"),
-                &[method.as_str()],
-                &content,
-            ));
-        }
-
-        let first = make_fingerprint_with_content(
-            "src/deploy.rs",
-            &["deploy_item"],
-            "fn deploy_item() {\n    scaffold_response();\n    read_request();\n    validate_component();\n    build_artifact();\n    upload_to_host();\n    run_post_hooks();\n    verify_release();\n    deploy_specific_step();\n}",
-        );
-        let second = make_fingerprint_with_content(
-            "src/upgrade.rs",
-            &["upgrade_item"],
-            "fn upgrade_item() {\n    scaffold_response();\n    read_request();\n    validate_component();\n    build_artifact();\n    upload_to_host();\n    run_post_hooks();\n    verify_release();\n    upgrade_specific_step();\n}",
-        );
-        fingerprints.push(first);
-        fingerprints.push(second);
-
-        let refs = fingerprints.iter().collect::<Vec<_>>();
-        let findings = detect_parallel_implementations(
-            &refs,
-            &std::collections::HashSet::new(),
-            &DuplicationDetectorConfig::default(),
-        );
-
-        assert_eq!(
-            findings.len(),
-            2,
-            "Domain-specific shared workflow calls should still flag after common scaffolding is discounted"
-        );
-        assert!(
-            findings
-                .iter()
-                .all(|f| f.description.contains("`build_artifact`")
-                    && !f.description.contains("`read_request`")),
-            "Findings should be driven by domain calls, got: {:?}",
-            findings.iter().map(|f| &f.description).collect::<Vec<_>>()
-        );
-    }
-}
+mod tests;
